add separate tx_body mod
