+++ rust/vote-tx-v2/src/gen_tx/decoding.rs
@@ -18,9 +18,6 @@ const CBOR_UUID_TAG: u64 = 37;
 /// `Vote` array struct length
 const VOTE_LEN: u64 = 3;
 
-/// `TxBody` array struct length
-const TX_BODY_LEN: u64 = 4;
-
 /// `GeneralizedTx` array struct length
 const GENERALIZED_TX_LEN: u64 = 2;
 
@@ -75,55 +72,6 @@ where
             .write_all(&sign_bytes)
             .map_err(minicbor::encode::Error::write)?;
 
-        Ok(())
-    }
-}
-
-impl<ChoiceT, ProofT, PropIdT> Decode<'_, ()> for TxBody<ChoiceT, ProofT, PropIdT>
-where
-    ChoiceT: for<'a> Cbor<'a>,
-    ProofT: for<'a> Cbor<'a>,
-    PropIdT: for<'a> Cbor<'a>,
-{
-    fn decode(d: &mut Decoder<'_>, (): &mut ()) -> Result<Self, minicbor::decode::Error> {
-        let Some(TX_BODY_LEN) = d.array()? else {
-            return Err(minicbor::decode::Error::message(format!(
-                "must be a defined sized array with {GENERALIZED_TX_LEN} entries"
-            )));
-        };
-
-        let vote_type = Uuid::decode(d, &mut ())?;
-        let event = EventMap::decode(d, &mut ())?;
-        let votes = Vec::<Vote<_, _, _>>::decode(d, &mut ())?;
-        if votes.is_empty() {
-            return Err(minicbor::decode::Error::message(
-                "votes array must has at least one entry",
-            ));
-        }
-        let voter_data = VoterData::decode(d, &mut ())?;
-        Ok(Self {
-            vote_type,
-            event,
-            votes,
-            voter_data,
-        })
-    }
-}
-
-impl<ChoiceT, ProofT, PropIdT> Encode<()> for TxBody<ChoiceT, ProofT, PropIdT>
-where
-    ChoiceT: for<'a> Cbor<'a>,
-    ProofT: for<'a> Cbor<'a>,
-    PropIdT: for<'a> Cbor<'a>,
-{
-    fn encode<W: minicbor::encode::Write>(
-        &self, e: &mut Encoder<W>, (): &mut (),
-    ) -> Result<(), minicbor::encode::Error<W::Error>> {
-        e.array(TX_BODY_LEN)?;
-        self.vote_type.encode(e, &mut ())?;
-        self.event.encode(e, &mut ())?;
-        self.votes.encode(e, &mut ())?;
-        self.voter_data.encode(e, &mut ())?;
         Ok(())
     }
 }
